backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Joran Timmerman <joranstimmerman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
alembic = "^1.9.1"
SQLAlchemy = "^1.4.46"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.0"
passlib = "^1.7.4"
pydantic = "^1.10.4"
jose = "^1.0.0"
python-multipart = "^0.0.5"
bcrypt = "^4.0.1"
PyJWT = "^2.6.0"
requests = "^2.28.2"
pandas = "^1.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
